# docs/openapi/gateway.yaml
openapi: 3.1.0
info:
  title: API Gateway
  description: |
    Centralized API Gateway handling cross-cutting concerns: authentication,
    rate limiting (60 req/min/key), CORS, correlation ID injection, security headers,
    and request routing to downstream services. Replicates AWS API Gateway behavior in dev.
  version: 1.0.0
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.legislation-rag.example.com
    description: Production (AWS API Gateway)
security:
  - ApiKeyAuth: []
paths:
...
  /v1/ingest:
    post:
      operationId: proxyIngest
      summary: Proxy to Ingestion Service document upload
      description: |
        Proxies to Ingestion Service POST /v1/ingest.
        Timeout: 60 seconds. Returns 504 if exceeded.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: "Document file (.txt, .md, .html, .pdf). Max size: 50 MB."
                chunking_strategy:
                  type: string
                  enum:
                    - fixed_size
                    - recursive
                    - semantic
                    - legal_hierarchical
                  description: Optional explicit chunking strategy
      responses:
        '201':
          description: Document ingested successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
              example:
                document_id: "doc-abc-123"
                filename: "transport_infrastructure_act_2024.md"
                chunks_produced: 42
                chunking_strategy: "legal_hierarchical"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /v1/documents:
    get:
      operationId: proxyListDocuments
      summary: Proxy to Ingestion Service document list
      description: Proxies to Ingestion Service GET /v1/documents.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Document list retrieved successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
              example:
                documents:
                  - document_id: "doc-abc-123"
                    filename: "transport_infrastructure_act_2024.md"
                    format: "md"
                    ingestion_date: "2024-01-15T10:30:00Z"
                    chunks_produced: 42
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (validated against configured key set)
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Optional correlation ID. If not provided, the gateway generates a UUID v4
        and injects it into the proxied request and response.
  headers:
    CorrelationId:
      schema:
        type: string
        format: uuid
      description: Correlation ID for distributed tracing (always present in responses)
...
  schemas:
...
    IngestResponse:
      type: object
      required:
        - document_id
        - filename
        - chunks_produced
        - chunking_strategy
      properties:
        document_id:
          type: string
        filename:
          type: string
        chunks_produced:
          type: integer
        chunking_strategy:
          type: string
          enum:
            - fixed_size
            - recursive
            - semantic
            - legal_hierarchical
    DocumentListResponse:
      type: object
      required:
        - documents
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
    DocumentSummary:
      type: object
      required:
        - document_id
        - filename
        - format
        - ingestion_date
        - chunks_produced
      properties:
        document_id:
          type: string
        filename:
          type: string
        format:
          type: string
        ingestion_date:
          type: string
          format: date-time
        chunks_produced:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
        - correlation_id
      properties:
        error_code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error description
        correlation_id:
          type: string
          format: uuid
          description: Request correlation ID
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "UNAUTHORIZED"
            message: "Missing or invalid X-API-Key header"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"
    Forbidden:
      description: Request blocked by security middleware (malicious pattern detected)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "FORBIDDEN"
            message: "Request blocked: potential SQL injection detected in request path"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"
    RateLimited:
      description: Rate limit exceeded (60 req/min/key)
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "RATE_LIMITED"
            message: "Rate limit exceeded: 60 requests per minute per API key"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "VALIDATION_ERROR"
            message: "Query must be between 1 and 2000 characters"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "INTERNAL_ERROR"
            message: "An unexpected error occurred"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"
    DependencyUnavailable:
      description: Downstream service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "SERVICE_UNAVAILABLE"
            message: "Downstream service is unreachable"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"
    GatewayTimeout:
      description: Proxy timeout exceeded (30s for queries, 60s for ingestion)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "GATEWAY_TIMEOUT"
            message: "Request to downstream service timed out after 30 seconds"
            correlation_id: "550e8400-e29b-41d4-a716-446655440000"